requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blake3==1.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorful==0.5.6
compressed-tensors==0.8.1
contourpy==1.3.1
cycler==0.12.1
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gradio==4.16.0
gradio_client==0.8.1
groq==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.4
httpx==0.24.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.8
lark==1.2.2
latex2mathml==3.77.0
libclang==18.1.1
linkify-it-py==2.0.3
llava @ git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd
lm-format-enforcer==0.10.9
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==2.1.5
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
namex==0.0.8
narwhals==1.19.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.58.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
peft==0.14.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyngrok==7.2.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
streamlit==1.41.1
svgwrite==1.4.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.15.1
toml==0.10.2
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.37.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
vllm==0.6.6.post1
watchfiles==1.0.3
wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
